Put transformation parsing and writing functions into separate files; transformation script almost successfully parses
